Moved filling teams from Sports to League
