+++ server/api/team/team.controller.js
@@ -74,89 +74,7 @@ exports.players = function(req, res) {
     });
 };
 
-exports.fill = function(req, res) {
-    var selectedSport = ""
-    var selectedTeams = []
-    var selectedPlayers = []
-
-    var user = req.user;
-    if (user.role !== 'admin') {
-        logger.log("error", "Filling teams is restricted to admins only");
-        return res.send (403);
-    }
-    Sport.find(req.params.sport_id).then(function(sport) {
-        return setSelectedSport(sport);
-    }).then(function() {
-        return Team.findAll({
-            where: [{ special: null, SportId: selectedSport.id, LeagueId: req.params.league_id }]
-        });
-    }).then(function(teams) {
-        return setSelectedTeams();
-    }).then(function() {
-        return Player.findAll({
-            where: { SportId: selectedSport.id }
-        })
-    }).then(function(players) {
-        return setSelectedPlayers(players);
-    }).then(function() {
-        return assignPlayers(setSelectedTeams, setSelectedPlayers);
-    }).then(function() {
-        return res.send(204);
-    }).catch (function(error) {
-        return handleError(res, error);
-    });
-};
-
 function handleError(res, error) {
     return res.send(500, error);
 }
 
-/**
-* Created by Sammy on 1/13/15
-**/
-
-function setSelectedSport(sport) {
-    selectedSport = sport;
-  }
-
-function setSelectedTeams(teams) {
-    selectedTeams = teams;
-  }
-
-function setSelectedPlayers(players) {
-    selectedPlayers = players;
-  }
-
-
-function assignPlayers(teams, players) {
-    var realWorldTeams = []
-
-    getRealWorldTeams(players);
-    return BPromise.map(realWorldTeams, function(realTeam) {
-        var realTeamIndex = realWorldTeams.indexOf(realTeam);
-        return db.Player.findAll({
-            where: { realWorldTeam: realTeam }
-        }).then(function(realTeamPlayers) {
-            return BPromise.map(realTeamPlayers, function(realTeamPlayer) {
-                if (realTeamIndex >= teams.length) {
-                    db.Team.find({
-                    where: {special: "freeagency"}
-                }).then(function(freeAgentTeam) {
-                    freeAgentTeam.addPlayer(realTeamPlayer);
-                    return freeAgentTeam.save();
-                });
-                } else {
-                    teams[realTeamIndex].addPlayer(realTeamPlayer);
-                    return teams[realTeamIndex].save();
-                }
-            });
-        });
-    });
-  }
-
-function getRealWorldTeams(players) {
-    realWorldTeams = _.chain(players)
-      .pluck('realWorldTeam')
-      .uniq()
-      .value();
-    }
