+++ server/api/team/team.spec.js
@@ -6,8 +6,6 @@ var request = require('supertest');
 var db = require('../models');
 var testUtil = require('../../components/test-util.js');
 
-var fillTest = require('../../components/fill_teams_test.js');
-
 db.sequelize.sync();
 
 describe('GET /api/teams', function() {
@@ -67,78 +65,4 @@ describe('GET /api/teams/:id/players', function() {
           done();
       });
   });
-});
-
-
-describe ('POST /api/teams/fill', function() {
-  var account1 = {
-    email: 'test1@test.com',
-    password: 'test',
-    role: 'admin'
-  };
-
-  var account2 = {
-    email: 'test2@test.com',
-    password: 'testing',
-    role: 'commish'
-  };
-
-  before(function(done) {
-    // Clear db before testing
-    var typesToClear = [
-          db.Sport,
-          db.User,
-          db.League,
-          db.Team,
-          db.Player,
-          db.PlayerAssignment
-      ];
-      testUtil.clearSequelizeTables(typesToClear,done);
-  });
-
-  before(function(done) {
-    db.User.create(account1).then(function(adminUser) {
-      return fillTest.fillTeams(adminUser);
-    }).then(function() {
-      done();
-    });
-  });
-
-  it('should have prepared 5 teams', function(done) {
-    db.Team.findAndCountAll().then(function(result) {
-      return result.count.should.equal(5);
-      done();
-    });
-  });
-
-  it('should have prepared a free agency team', function(done) {
-    db.Team.find({where: {special: "freeagency"}
-    }).then(function () {
-      done();
-    })
-  })
-
-  it('should have attached players to teams', function(done) {
-    db.PlayerAssignment.find( {where: {"TeamId" : 1} }).then(function(result) {
-      return result.PlayerId.should.equal(1);
-      done();
-    });
-  });
-
-  it('should have added 4 players to free agency', function(done) {
-    db.PlayerAssignment.findAndCountAll( {where: {"TeamId" : 4} }).then(function(result) {
-      return result.count.should.equal(4);
-      done();
-    });
-  });
-
-  it('should only allow adminUser to send and complete the request', function(done) {
-    db.User.create(account2).then(function(commishUser) {
-      return fillTest.fillTeams(commishUser);
-    }).error(function (error) {
-      console.log(error);
-      done();
-    });
-  });
-
 });
