+++ server/logger/undefined.log
@@ -2,3 +2,7 @@
 {"level":"info","message":"Suceeded in populating league","timestamp":"2015-01-28T06:53:35.786Z"}
 {"level":"info","message":"Suceeded in populating league","timestamp":"2015-01-29T01:14:44.316Z"}
 {"level":"info","message":"Suceeded in populating league","timestamp":"2015-02-03T11:55:21.505Z"}
+{"level":"info","message":"Suceeded in populating league","timestamp":"2015-02-04T23:48:29.317Z"}
+{"level":"info","message":"Suceeded in populating league","timestamp":"2015-02-04T23:50:19.439Z"}
+{"level":"info","message":"Suceeded in populating league","timestamp":"2015-02-05T06:51:28.559Z"}
+{"level":"info","message":"Suceeded in populating league","timestamp":"2015-02-05T07:46:19.060Z"}
